requirements.txt
pyeit.egg==info
matplotlib==3.3.4
numpy==1.20.3
pandas==1.2.4
PyQt5==5.15.4
pyserial==3.5
